requirements.txt
torch>=2.1.0
datasets>=2.16.0
transformers>=4.36.0
tokenizers>=0.15.0
pyyaml>=6.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
tqdm>=4.66.0
tensorboard>=2.15.0
pytest>=7.4.0
